Setup KMS key on bootstrap
One of the first things that needs to be done when setting up
CI for an AWS account is to add secrets to the SSM.  To add
secrets to the SSM you need a KMS key.  We add a KMS key to
boostrap so we can add secrets and run sceptre to retrieve
them and inject them into the CF templates.